pyproject.toml
[project]
name = "drift-e-commerce"
version = "0.1.0"
description = "E-Commerce Website"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "pymongo>=4.15.5",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.10",
]
[project.scripts]
website = "main:main"
